ADD/ validation to mode-shipment service

+++ Back/src/mode-shipment/mode-shipment.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { ModeShipmentRepository } from './mode-shipment.repository';
 import { v4 as uuid } from 'uuid';
 import { ModeShipmentDto, ModeShipmentParcialDto } from './mode-shipment.dto';
@@ -9,22 +9,31 @@ export class ModeShipmentService {
 
   async getAll() {
     const MS = await this.modeShipmentRepository.getAll();
-    return MS;
+    if (!MS || MS.length == 0)
+      throw new BadRequestException('There are no mode shipment');
+    else return MS;
   }
 
   async getById(id: uuid) {
+    if (!id) throw new BadRequestException('Id must be defined');
     const MS = await this.modeShipmentRepository.getById(id);
-    return MS;
+    if (!MS) throw new BadRequestException('Mode shipment not found');
+    else return MS;
   }
 
   async getByName(name: string) {
+    if (!name || name.length == 0)
+      throw new BadRequestException('Name must be defined');
     const MS = await this.modeShipmentRepository.getByName(name);
-    return MS;
+    if (!MS) throw new BadRequestException('Mode shipment not found');
+    else return MS;
   }
 
   async getByPrice(price: number) {
+    if (!price) throw new BadRequestException('Price must be defined');
     const MS = await this.modeShipmentRepository.getByPrice(price);
-    return MS;
+    if (!MS) throw new BadRequestException('Mode shipment not found');
+    else return MS;
   }
 
   async create(modeShipment: ModeShipmentDto) {
@@ -32,10 +41,16 @@ export class ModeShipmentService {
   }
 
   async update(id: uuid, modeShipment: ModeShipmentParcialDto) {
+    if (!id) throw new BadRequestException('Id must be defined');
+    if (!this.getById(id))
+      throw new BadRequestException('Mode shipment not found');
     return await this.modeShipmentRepository.update(id, modeShipment);
   }
 
   async delete(id: uuid) {
+    if (!id) throw new BadRequestException('Id must be defined');
+    if (!this.getById(id))
+      throw new BadRequestException('Mode shipment not found');
     return await this.modeShipmentRepository.delete(id);
   }
 }
